Reduce the memory consumption of our abstract dex instructions.
R=herhut@google.com

+++ src/main/java/com/android/tools/r8/code/Format10t.java
@@ -10,7 +10,7 @@ import java.nio.ShortBuffer;
 
 abstract class Format10t extends Base1Format {
 
-  public final /* offset */ int AA;
+  public final /* offset */ byte AA;
 
   // +AA | op
   Format10t(int high, BytecodeStream stream) {
@@ -21,7 +21,7 @@ abstract class Format10t extends Base1Format {
 
   protected Format10t(int AA) {
     assert Byte.MIN_VALUE <= AA && AA <= Byte.MAX_VALUE;
-    this.AA = AA;
+    this.AA = (byte) AA;
   }
 
   public void write(ShortBuffer dest, ObjectToOffsetMapping mapping) {
