+++ src/main/java/com/android/tools/r8/code/Format20t.java
@@ -10,7 +10,7 @@ import java.nio.ShortBuffer;
 
 abstract class Format20t extends Base2Format {
 
-  public final /* offset */ int AAAA;
+  public final /* offset */ short AAAA;
 
   // øø | op | +AAAA
   Format20t(int high, BytecodeStream stream) {
@@ -20,7 +20,7 @@ abstract class Format20t extends Base2Format {
 
   protected Format20t(int AAAA) {
     assert Short.MIN_VALUE <= AAAA && AAAA <= Short.MAX_VALUE;
-    this.AAAA = AAAA;
+    this.AAAA = (short) AAAA;
   }
 
   public void write(ShortBuffer dest, ObjectToOffsetMapping mapping) {
